pfe-backend/pyproject.toml
[project]
name = "pfe-backend"
version = "0.1.0"
description = "PFE backend"
authors = [
    {name = "rayhan"}
]
requires-python = "^3.11"
dependencies = [
    "requests (>=2.33.1,<3.0.0)",
    "pydantic (>=2.13.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "mcp (>=1.27.0,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "elasticsearch (<9.0.0)",
    "anthropic (>=0.50.0)",
    "openai (>=1.50.0)",
    "fastembed (>=0.4.0,<1.0.0)",
    "rank-bm25 (>=0.2.2)",
    "numpy (>=1.24.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "duckduckgo-search (>=6.0,<7.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "pypdf (>=6.10.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
